// packages/beecli/sdk/beetrade-v2/models/CreateTradingClientRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateTradingClientRequest
 */
export interface CreateTradingClientRequest {
    /**
     * 
     * @type {Date}
     * @memberof CreateTradingClientRequest
     */
    dateOfBirth?: Date;
    /**
     * 
     * @type {string}
     * @memberof CreateTradingClientRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateTradingClientRequest
     */
    name: string;
}

/**
 * Check if a given object implements the CreateTradingClientRequest interface.
 */
export function instanceOfCreateTradingClientRequest(value: object): value is CreateTradingClientRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateTradingClientRequestFromJSON(json: any): CreateTradingClientRequest {
    return CreateTradingClientRequestFromJSONTyped(json, false);
}

export function CreateTradingClientRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateTradingClientRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'dateOfBirth': json['dateOfBirth'] == null ? undefined : (new Date(json['dateOfBirth'])),
        'description': json['description'] == null ? undefined : json['description'],
        'name': json['name'],
    };
}

export function CreateTradingClientRequestToJSON(json: any): CreateTradingClientRequest {
    return CreateTradingClientRequestToJSONTyped(json, false);
}

export function CreateTradingClientRequestToJSONTyped(value?: CreateTradingClientRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dateOfBirth': value['dateOfBirth'] == null ? value['dateOfBirth'] : value['dateOfBirth'].toISOString(),
        'description': value['description'],
        'name': value['name'],
    };
}