// packages/beecli/sdk/beetrade-v2/models/UpdateTradingClientRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Beetrade API
 * API documentation for Beetrade
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateTradingClientRequest
 */
export interface UpdateTradingClientRequest {
    /**
     * 
     * @type {Date}
     * @memberof UpdateTradingClientRequest
     */
    dateOfBirth?: Date;
    /**
     * 
     * @type {string}
     * @memberof UpdateTradingClientRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateTradingClientRequest
     */
    name?: string;
}

/**
 * Check if a given object implements the UpdateTradingClientRequest interface.
 */
export function instanceOfUpdateTradingClientRequest(value: object): value is UpdateTradingClientRequest {
    return true;
}

export function UpdateTradingClientRequestFromJSON(json: any): UpdateTradingClientRequest {
    return UpdateTradingClientRequestFromJSONTyped(json, false);
}

export function UpdateTradingClientRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateTradingClientRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'dateOfBirth': json['dateOfBirth'] == null ? undefined : (new Date(json['dateOfBirth'])),
        'description': json['description'] == null ? undefined : json['description'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function UpdateTradingClientRequestToJSON(json: any): UpdateTradingClientRequest {
    return UpdateTradingClientRequestToJSONTyped(json, false);
}

export function UpdateTradingClientRequestToJSONTyped(value?: UpdateTradingClientRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dateOfBirth': value['dateOfBirth'] == null ? value['dateOfBirth'] : value['dateOfBirth'].toISOString(),
        'description': value['description'],
        'name': value['name'],
    };
}